feat: add support for .env files
This adds a new class Environment that combines system environment
variables with those found in a .env file. It replaces the direct
use of System.getenv in Configurator by using a custom
StringSubstitutor StringLookup class that uses the new Environment
class.

This is an improvement over the previous attempt to add .env file
support. It takes into account StringSubstitutor's default value
behavior.